consumer/requirements.txt
kafka-python==2.0.2
river==0.21.0
numpy==1.24.0
